# public/openapi.yaml
openapi: "3.0.3"
info:
  title: 競売物件データAPI
  version: "1.0"
  description: |
    裁判所の競売物件・訳あり激安物件データを毎日自動収集し、REST APIで提供します。
    無料プランは1日100リクエスト。APIキーはhttps://auction-property-api.vercel.appから取得。
  contact:
    url: https://auction-property-api.vercel.app/docs
servers:
  - url: https://auction-property-api.vercel.app/api/v1
    description: 本番環境
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    AuctionProperty:
      type: object
      properties:
        id:
          type: string
          example: "court_13_R6_ke_1234"
        court:
          type: string
          example: "東京地方裁判所"
        case_number:
          type: string
          example: "令6(ケ)第1234号"
        property_type:
          type: string
          enum: [land, building, apartment, farm]
        address:
          type: string
          example: "東京都渋谷区代々木1-1-1"
        building_name:
          type: string
          nullable: true
        area_sqm:
          type: number
          nullable: true
        floor:
          type: integer
          nullable: true
        total_floors:
          type: integer
          nullable: true
        base_price:
          type: integer
          example: 12000000
        auction_start_date:
          type: string
          format: date
          nullable: true
        auction_end_date:
          type: string
          format: date
          example: "2024-04-22"
        status:
          type: string
          enum: [open, sold, cancelled]
        court_url:
          type: string
          format: uri
        lat:
          type: number
          nullable: true
        lon:
          type: number
          nullable: true
        scraped_at:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        docs_url:
          type: string
paths:
  /auctions:
    get:
      summary: 競売物件一覧
      description: 裁判所の競売物件一覧を取得します。認証必須。
      operationId: listAuctions
      parameters:
        - name: prefecture
          in: query
          schema:
            type: string
          description: "都道府県コード（2桁）例: 13"
        - name: category
          in: query
          schema:
            type: string
            enum: [land, building, apartment, farm]
          description: "物件種別"
        - name: min_price
          in: query
          schema:
            type: integer
          description: "最低売却基準額（円）"
        - name: max_price
          in: query
          schema:
            type: integer
          description: "最高売却基準額（円）"
        - name: status
          in: query
          schema:
            type: string
            enum: [open, sold, cancelled]
            default: open
          description: "状態（デフォルト: open）"
        - name: limit
          in: query
          schema:
            type: integer
            maximum: 100
            default: 20
          description: "取得件数"
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: "ページング開始位置"
      responses:
        "200":
          description: 物件一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuctionProperty"
        "401":
          description: APIキー不正
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: レート制限超過
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auctions/stats:
    get:
      summary: 統計情報（認証不要）
      description: 現在公開中の競売物件数を物件種別ごとに返します。認証不要。
      operationId: getStats
      security: []
      responses:
        "200":
          description: 件数統計
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_open:
                    type: integer
                  by_type:
                    type: object
                  updated_at:
                    type: string
                    format: date-time
  /auctions/{id}:
    get:
      summary: 競売物件詳細
      description: IDを指定して競売物件の詳細情報を取得します。
      operationId: getAuction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "物件ID（例: court_13_R6_ke_1234）"
      responses:
        "200":
          description: 物件詳細
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuctionProperty"
        "401":
          description: APIキー不正
        "404":
          description: 物件が見つかりません
  /webhooks:
    post:
      summary: Webhook登録（Proプラン以上）
      description: 条件に合致する新着競売物件が追加された際に通知するWebhookを登録します。
      operationId: createWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint_url]
              properties:
                endpoint_url:
                  type: string
                  format: uri
                  description: "通知先URL（HTTPSのみ）"
                filter_prefecture:
                  type: string
                  description: "都道府県フィルター"
                filter_property_type:
                  type: string
                  enum: [land, building, apartment, farm]
                filter_min_price:
                  type: integer
                filter_max_price:
                  type: integer
      responses:
        "201":
          description: Webhook登録完了
        "401":
          description: APIキー不正
        "403":
          description: プランがPro以上でない